add no-event-view and refactoring
