+++ src/utils/event.js
@@ -9,11 +9,6 @@ const DATE_FORMAT = 'MMM D';
 const TIME_FORMAT = 'hh:mm';
 const DATE_FORMAT_FULL = 'DD/MM/YY hh:mm';
 
-
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
-
 function formatDate(dueDate) {
   return dueDate ? dayjs(dueDate).format(DATE_FORMAT) : '';
 }
@@ -33,4 +28,4 @@ function differenceTime(toTime, fromTime) {
   return diffMin < 60 ? `${diffMin + 1}m` : `${diffHour}h ${diffResult + 1}m`;
 }
 
-export {getRandomArrayElement, formatDate, formatTime, differenceTime, formatDateFull};
+export {formatDate, formatTime, differenceTime, formatDateFull};
